feat(auth): adiciona auto refresh no token
Atualiza automaticamente o acess_token
