+++ app/(privated)/cadastrarUsuario/services/cadastrarUsuario.tsx
@@ -1,5 +1,6 @@
 'use server'
 import { authOptions } from "@/app/api/auth/[...nextauth]/route"
+import { BACKEND_URL } from "@/app/constants";
 import { getServerSession } from "next-auth"
 
 /*
@@ -9,12 +10,12 @@ import { getServerSession } from "next-auth"
 export async function cadastrarUsuario(body: {}) {
 	const session = await getServerSession(authOptions);
 
-	const request = await fetch('http://localhost:5000/users', {
+	const request = await fetch(BACKEND_URL + '/users/', {
 		method: 'POST',
 		headers: {
 			'Accept': 'application/json',
 			'Content-Type': 'application/json',
-			'Authorization': `Bearer ` + session?.backendTokens.access_token
+			authorization: `Bearer ` + session?.backendTokens.access_token
 		},
 		body: JSON.stringify(body)
 	})
