--- app/cadastrarSala/services/cadastrarSalaFisica.tsx
@@ -1,18 +1,0 @@
-
-/*
- * envia o body para cadastro do usuário no backend
- *
- * */
-export async function cadastrarSala(body: {}) {
-
-    const request = fetch('http://localhost:5000/physicalrooms', {
-      method: 'POST',
-      headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(body)
-    })
-    const response = await request
-    return response.json()
-  }
