+++ app/components/Navbar/NavbarMenu.tsx
@@ -21,7 +21,7 @@ export default function NavbarMenu() {
                     <Link href={'/'}><NavbarOptionsMenu hasCorFundo={true} texto='Tela Inicial' /></Link>
                     <Link href={'/agendar'}><NavbarOptionsMenu hasCorFundo={false} texto='Agendar' /></Link>
                     <Link href={'/meusAgendamentos'} > <NavbarOptionsMenu hasCorFundo={true} texto='Meus Agendamentos' /></Link>
-                    <Link href={''} > <NavbarOptionsMenu hasCorFundo={false} texto='Cadastrar Sala' /></Link>
+                    <Link href={'/cadastrarSala'} > <NavbarOptionsMenu hasCorFundo={false} texto='Cadastrar Sala' /></Link>
                     <Link href={'/cadastrarUsuario'}><NavbarOptionsMenu hasCorFundo={true} texto='Cadastrar Usuário' /></Link>
                     <Link href={''} ><NavbarOptionsMenu hasCorFundo={false} texto='Visualizar Salas' /></Link>
                     <Link href={''} ><NavbarOptionsMenu hasCorFundo={true} texto='Visualizar Usuários' /></Link>
