+++ app/constants.ts
@@ -1,0 +1,3 @@
+export const BACKEND_URL = "http://backend:5000" // Docker users
+
+// export const BACKEND_URL = "http://backend:5000" // no Docker users
